Button to save a row as template.
